+++ openfisca_core/periods/tests/test_instant.py
@@ -1,3 +1,5 @@
+import datetime
+
 import pytest
 
 from openfisca_core import periods
@@ -6,6 +8,29 @@ from openfisca_core import periods
 @pytest.fixture
 def instant():
     return periods.Instant((2020, 2, 29))
+
+
+@pytest.mark.parametrize("arg, expected", [
+    [None, None],
+    [periods.Instant((1, 1, 1)), datetime.date(1, 1, 1)],
+    [periods.Instant((4, 2, 29)), datetime.date(4, 2, 29)],
+    [(1, 1, 1), datetime.date(1, 1, 1)],
+    ])
+def test_to_date(arg, expected):
+    assert periods.Instant.to_date(arg) == expected
+
+
+@pytest.mark.parametrize("arg, error", [
+    [periods.Instant((-1, 1, 1)), ValueError],
+    [periods.Instant((1, -1, 1)), ValueError],
+    [periods.Instant((1, 1, -1)), ValueError],
+    [periods.Instant((1, 13, 1)), ValueError],
+    [periods.Instant((1, 1, 32)), ValueError],
+    [periods.Instant((1, 2, 29)), ValueError],
+    ])
+def test_to_date_with_an_invalid_argument(arg, error):
+    with pytest.raises(error):
+        periods.Instant.to_date(arg)
 
 
 @pytest.mark.parametrize("offset, unit, expected", [